+++ LJC.Com.LogService/LJC.Com.LogViewWeb/LogsNew/ReadLog.ashx.cs
@@ -34,8 +34,8 @@ namespace LJC.Com.LogViewWeb.Scripts.Pages.LogsNew
                 {
                     end = DateTime.Now;
                 }
-                var pos = -1;
-                if (!int.TryParse(context.Request["pos"], out pos))
+                var pos = -1L;
+                if (!long.TryParse(context.Request["pos"], out pos))
                 {
                     pos = 0;
                 }
